minor internal refactoring
