+++ src/app/shared/models/GraphUtil.model.ts
@@ -1,0 +1,11 @@
+export interface Node {
+    id: string;
+    x: number;
+    y: number;
+}
+
+export interface Coordinates {
+    x: number;
+    y: number;
+}
+
